refactor: main.h->minirt.h변경 및 인클루드 헤더 위치 변경 리팩토링

+++ srcs/parse/get_next_line_utils.c
@@ -1,6 +1,6 @@
-#include "../../include/main.h"
+#include "../../include/get_next_line.h"
 
-size_t	ft_strlen2(const char *str)
+int	ft_strlen2(const char *str)
 {
 	int	i;
